fix ia32 build error. With IA32 tool chain, we can not use a 64bit data to divide a 32 bit data. MS tool chain may introduce an intrinsic function.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@10267 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Bus/Ata/AtaBusDxe/AtaPassThruExecute.c
@@ -304,8 +304,8 @@ IdentifyAtaDevice (
       // Check lowest alignment of logical blocks within physical block
       //
       if ((IdentifyData->alignment_logic_in_phy_blocks & (BIT14 | BIT15)) == BIT14) {
-        BlockMedia->LowestAlignedLba = (EFI_LBA) (BlockMedia->LogicalBlocksPerPhysicalBlock - (IdentifyData->alignment_logic_in_phy_blocks & 0x3fff)) %
-          BlockMedia->LogicalBlocksPerPhysicalBlock;
+        BlockMedia->LowestAlignedLba = (EFI_LBA) ((BlockMedia->LogicalBlocksPerPhysicalBlock - ((UINT32)IdentifyData->alignment_logic_in_phy_blocks & 0x3fff)) %
+          BlockMedia->LogicalBlocksPerPhysicalBlock);
       }
     }
     //
